changed types to floats

+++ source/CNCVibration/CncMeasurement.Core/SharedModels.cs
@@ -36,8 +36,8 @@ namespace CncMeasurement.Core.models
     public class MeasurementConfig
     {
         public string ChannelName { get; set; } //= "cDAQ1Mod1/ai0";
-        public double SampleRate { get; set; } //= 10240.0;
-        public double DurationSeconds { get; set; } //= 2.0;
+        public float SampleRate { get; set; } //= 10240.0;
+        public float DurationSeconds { get; set; } //= 2.0;
         public int SpindleRpm { get; set; }
     }
